pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "webshell-detector"
version = "2.0.0"
description = "Machine Learning based webshell detector"
authors = [
    { name="Nikhil Kumar" }
]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0"
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312", "py313"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | build
  | dist
  | Output
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    ".git",
    "venv",
    ".venv",
    "Output",
    "__pycache__"
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
]
ignore = [
    "E501", # line too long (handled by black)
    "W191", # indentation contains tabs
    "B008", # FastAPI File/Form defaults
    "B904", # raise from
    "B905", # zip strict
    "B007", # Unused loop var
    "F841", # Unused local var
    "B035", # Dict comprehension static key
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
